Summary
The gazebo commit contains files to commence and consequently continue
work on the gazebo module of the Udacity course: Robotics Software
Engineering. The given files are:
    - The world file
    - The robot model file
    - The "Hello World" C++ plugin script

This first iteration lunches a gazebo world simply populated by the robot model
included and runs a script at startup that prints hello world to the shell.